+++ src/components/App/AppMain/App.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useState } from 'react';
+import { FC } from 'react';
 import 'bulma/bulma.sass';
 import '@fortawesome/fontawesome-free/css/all.css';
 import './App.scss';
@@ -7,57 +7,19 @@ import classNames from 'classnames';
 import { PostsList } from '../../Posts/PostList/PostsList';
 import { PostDetails } from '../../Posts/PostDetails/PostDetails';
 import { UserSelector } from '../../Users/UserSelector/UserSelector';
-import { Post } from '../../../types/Post';
-import { User } from '../../../types/User';
-import { getUsers } from '../../../api/users';
-import { getPosts } from '../../../api/posts';
-import { ErrorType } from '../../../types/ErrorType';
 import { Error } from '../AppError/Error';
+import { useApp } from '../hooks/useApp';
 
-export const App: React.FC = () => {
-  const [posts, setPosts] = useState<Post[]>([]);
-  const [users, setUsers] = useState<User[]>([]);
-  const [error, setError] = useState<ErrorType>(ErrorType.None);
-  const [selectedPost, setSelectedPost] = useState<Post | null>(null);
-  const [selectedUser, setSelectedUser] = useState<User | null>(null);
-
-  const loadUsers = useCallback(async () => {
-    try {
-      const usersFromServer = await getUsers();
-
-      setUsers(usersFromServer);
-    } catch {
-      setError(ErrorType.LoadUsers);
-    }
-  }, []);
-
-  const loadPosts = useCallback(async (userId: number) => {
-    try {
-      const postsFromServer = await getPosts(userId);
-
-      setPosts(postsFromServer);
-    } catch {
-      setError(ErrorType.LoadPosts);
-    }
-  }, []);
-
-  const selectingPost = useCallback(async (user: User) => {
-    setSelectedUser(user);
-  }, []);
-
-  useEffect(() => {
-    loadUsers();
-  }, []);
-
-  useEffect(() => {
-    setSelectedPost(null);
-
-    if (selectedUser) {
-      loadPosts(selectedUser.id);
-    } else {
-      setPosts([]);
-    }
-  }, [selectedUser?.id]);
+export const App: FC = () => {
+  const {
+    posts,
+    users,
+    error,
+    selectedPost,
+    selectedUser,
+    selectingPost,
+    setSelectedPost,
+  } = useApp();
 
   return (
     <main className="section">
