+++ src/types/ErrorType.ts
@@ -3,7 +3,7 @@ export enum ErrorType {
   LoadComments = 'Unable to load a comments',
   LoadPosts = 'Unable to load a posts',
   LoadAddComments = 'Unable to add a comment',
-  LoadingButton = 'Loading',
+  LoadingButton = 'Loading', // delete
   InputNameError = 'Name is required',
   InputEmailError = 'Email is required',
   InputBodyError = 'Enter some text',
